+++ examples/polymer-cli/.storybook/addons.js
@@ -2,3 +2,4 @@ import '@storybook/addon-actions/register';
 import '@storybook/addon-notes/register';
 import '@storybook/addon-knobs/register';
 import '@storybook/addon-viewport/register';
+import '@storybook/addon-options/register';
