+++ examples/polymer-cli/package.json
@@ -11,6 +11,7 @@
     "@storybook/addon-actions": "^3.4.0-alpha.7",
     "@storybook/addon-knobs": "^3.4.0-alpha.7",
     "@storybook/addon-notes": "^3.4.0-alpha.7",
+    "@storybook/addon-options": "^3.4.0-alpha.7",
     "@storybook/addon-viewport": "^3.4.0-alpha.7",
     "@storybook/polymer": "^3.4.0-alpha.7",
     "@webcomponents/webcomponentsjs": "^1.1.0",
